Resize grid with function parameters
Grid logic was replaced into createGrid, making possible to manipulate the grid size with the squareNum parameter, which is set in the addEventListener.

The addEventListener below createGrid remove the previous grid with a while loop to create another one based in the value received in the prompt. Then the createGrid function is called again with the prompt value as an argument.